CRUD UserProfile

+++ server/Controller/ControllerUserProfile.js
@@ -8,6 +8,14 @@ class ControllerUserProfile {
     save() {
         let UserProfile = require("../Models/UserProfile")
 
+        let fieldsdEmpty = this.checkFields(this.name, this.email, this.password)
+
+        if(fieldsdEmpty) {
+            return new Promise((resolve, reject) => {
+                reject(fieldsdEmpty)
+            })
+        }
+
         UserProfile = new UserProfile({
             name:this.name,
             email:this.email,
@@ -22,6 +30,14 @@ class ControllerUserProfile {
     update(id) {
         let UserProfile = require("../Models/UserProfile")
 
+        let fieldsdEmpty = this.checkFields(this.name, this.email, this.password)
+
+        if(fieldsdEmpty) {
+            return new Promise((resolve, reject) => {
+                reject(fieldsdEmpty)
+            })
+        }
+
         return UserProfile.updateOne({_id:id}, {$set:{name:this.name, email:this.email, password:this.password}}).then(doc => {
             return doc
         })
@@ -30,9 +46,21 @@ class ControllerUserProfile {
     delete(id) {
         let UserProfile = require("../Models/UserProfile")
 
+        if(id.length != 24) {
+            return new Promise((resolve, reject) => {
+                reject("Informe um id para deletar !!")
+            })
+        }
+
         return UserProfile.deleteOne({_id:id}).then(doc => {
             return doc
         })
+    }
+
+    checkFields(name, email, password) {
+        if(!name || !email || !password) {
+           return "Não podem haver campos vazios !!!"
+        }
     }
 }
